Add python3 compatibility. Still compatible with python 2.6.
